display/requirements.txt
gps==3.19
luma.oled==3.4.0
